Feat: signal함수 처리
1. 명령어를 받기전과 받은 후의 처리를 한 함수로 묶어서 처리하도록 변경

resolved #98
