+++ Makefile
@@ -6,7 +6,7 @@
 #    By: sehjang <sehjang@student.42seoul.k>        +#+  +:+       +#+         #
 #                                                 +#+#+#+#+#+   +#+            #
 #    Created: 2022/07/20 15:20:00 by sehjang           #+#    #+#              #
-#    Updated: 2022/09/13 23:07:35 by sunwchoi         ###   ########.fr        #
+#    Updated: 2022/09/21 09:38:09 by sunwchoi         ###   ########.fr        #
 #                                                                              #
 # **************************************************************************** #
 
@@ -17,7 +17,8 @@ NAME = minishell
 
 SRCS = 
 
-SRCS_MANDATORY = srcs/setting.c \
+SRCS_MANDATORY = srcs/signal.c \
+				 srcs/setting.c \
 				 srcs/main.c \
 				 srcs/parsing.c \
 				 srcs/parsing_utils.c \
