+++ includes/setting.h
@@ -6,7 +6,7 @@
 /*   By: sehjang <sehjang@student.42.fr>            +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/09/16 18:59:22 by sehjang           #+#    #+#             */
-/*   Updated: 2022/09/16 18:59:23 by sehjang          ###   ########.fr       */
+/*   Updated: 2022/09/21 09:37:37 by sunwchoi         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -22,8 +22,6 @@ typedef struct s_ev
 	t_list	*evl;
 }	t_ev;
 
-void	main_handler(int sig);
-void	sub1_handler(int sig);
 void	set_terminal(void);
 void	ret_terminal(void);
 void	get_ev(t_ev *ev, char **envp);
