+++ includes/signal.h
@@ -1,0 +1,6 @@
+#ifndef SIGNAL_H
+# define SIGNAL_H
+
+void	signal_process_in_waiting();
+void	signal_process_in_command();
+#endif
