+++ srcs/exec_cmd.c
@@ -6,7 +6,7 @@
 /*   By: sehjang <sehjang@student.42.fr>            +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/09/16 18:57:01 by sehjang           #+#    #+#             */
-/*   Updated: 2022/09/16 18:57:02 by sehjang          ###   ########.fr       */
+/*   Updated: 2022/09/21 09:40:13 by sunwchoi         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
